% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matches.R
\name{find_matches}
\alias{find_matches}
\title{Find matches in regulation of proteins and metabolites against a database}
\usage{
find_matches(db, prot_reg, metab_reg)
}
\arguments{
\item{db}{Database against which to compare protein and metabolite regulation data}

\item{prot_reg}{Protein regulation vector with names. Values should be < 0 for downregulated elements and > 0 for upregulated elements.}

\item{metab_reg}{Metabolite regulation vector. Values should be < 0 for downregulated elements and > 0 for upregulated elements.}
}
\value{
A list of matching results, each list element corresponds to a single protein. Each list item contains elements:
  \item{consumed}{consumed metabolites in \code{db} that are downregulated in \code{metab_reg}}
  \item{produced}{produced metabolites in \code{db} that are upregulated in \code{metab_reg}}
  \item{nondir_down}{non-directional metabolites in \code{db} that are downregulated in \code{metab_reg}}
  \item{nondir_up}{non-directional metabolites in \code{db} that are upregulated in \code{metab_reg}}
Note that non-directional metabolites are those that are listed as being both consumed and produced in the database.
}
\description{
[TODO: Insert more information about the relevance of this]
}
