% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_data_documentation.R
\docType{data}
\name{reactome_data_version}
\alias{reactome_data_version}
\title{Reactome version used to produce data objects for mapping}
\format{
An object of class \code{character} of length 1.
}
\source{
\url{https://reactome.org}
}
\usage{
reactome_data_version
}
\description{
A string containing the version of the Reactome data that was parsed to form the
\code{reactome_mapping_db} and \code{reactome_reactions} objects.
}
\seealso{
\code{\link{reactome_mapping_db}}, \code{\link{reactome_reactions}}
}
\keyword{datasets}
