% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_data_documentation.R
\docType{data}
\name{reactome_mapping_db}
\alias{reactome_mapping_db}
\title{Mapping from proteins to reactions and metabolites}
\format{
A \code{tibble} with the following fields:
\describe{
  \item{protein}{Uniprot ID (character string)}
  \item{reaction_id}{reaction ID from Reactome}
  \item{consumed}{list of double-valued CHEBI IDs of metabolites consumed in the corresponding reaction}
  \item{produced}{list of double-valued CHEBI IDs of metabolites produced in the corresponding reaction}
}
}
\source{
\url{https://reactome.org}
}
\usage{
reactome_mapping_db
}
\description{
A tibble that maps from proteins to reactions and the metabolites consumed and produced
from those reactions. The proteins are Uniprot IDs extracted from the reaction 
\code{modifier} field in the Reactome SBML source files. Metabolites are CHEBI IDs extracted
from the reaction \code{consumed} and \code{produced} fields of the SBML files.
}
\seealso{
\code{\link{reactome_reactions}}, \code{\link{reactome_data_version}}
}
\keyword{datasets}
